Move %LIBRARY_LIB%/lib/cmake.
The FindBoost.cmake module will defer to the cmake files left in
%LIBRARY_LIB%/lib/cmake when building anything that uses cmake that also
links to boost. These boost cmake config files will force the build of
that dependent project to use static libraries. If the cmake config
files are not in this standard place, FindBoost.cmake will do the right
thing regarding the use of shared libraries if the depedent project
wants that.